+++ CHANGELOG.md
@@ -13,6 +13,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
+## [Unreleased] - 2023-04-26
+
+### Added
+- BaseConsumer
+- InboxConsumer
+- Instrumentation listeners: sentry, logger, yabeda
+
+### Changed
+
+### Fixed
+
 ## [Unreleased]
 
 ## [0.1.0] - 2023-04-19